requirements.txt
Django>=4.2,<5.0
requests>=2.28.0
python-dotenv>=0.19.0

